Split FileManager into Platform Dependent Files
